+++ filen-cli/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - autocompletion in interactive mode, including remote paths
 - interactive `search` command to globally search for files and directories in the drive
 - `upload` and `download` commands to transfer files and directories to and from the drive
+- `trash-restore` and `trash-delete` commands to restore or permanently delete files and directories from the trash
 
 ### Changed